requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Security & Authentication
bcrypt==4.2.1
PyJWT==2.9.0
python-jose[cryptography]==3.3.0

# Google Gemini AI
google-generativeai==0.8.3

# File handling
python-multipart==0.0.20

# Environment variables
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-magic==0.4.27

# Excel/CSV
pandas==2.1.4
openpyxl==3.1.2

# Utilidades
python-dateutil==2.8.2
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1